+++ tools/lib/canmore_cpp_core/include/canmore_cpp/RegMappedClient.hpp
@@ -92,7 +92,7 @@ public:
 
     uint32_t readRegister(uint8_t mode, uint8_t page, uint8_t offset);
     void writeRegister(uint8_t mode, uint8_t page, uint8_t offset, uint32_t data);
-    void readArray(uint8_t mode, uint8_t page, uint8_t offsetStart, std::vector<uint32_t> &dst, uint8_t numWords);
+    void readArray(uint8_t mode, uint8_t page, uint8_t offsetStart, std::vector<uint32_t> &dst, unsigned int numWords);
     void writeArray(uint8_t mode, uint8_t page, uint8_t offsetStart, std::vector<uint32_t> &data);
     void writeStringPage(uint8_t mode, uint8_t page, const std::string &data);
     std::string readStringPage(uint8_t mode, uint8_t page);
